First approach to upgrading 1.9
