+++ valuenetwork/templates/account/base.html
@@ -1,6 +1,6 @@
 {% extends "subnav_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 
 {% block body_class %}account{% endblock %}
