+++ valuenetwork/templates/account/email_confirmed.html
@@ -1,6 +1,6 @@
 {% extends "site_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 
 {% block head_title %}{% trans "Email confirmed" %}{% endblock %}
