+++ valuenetwork/templates/account/login.html
@@ -1,4 +1,4 @@
-{% load url from future %}
+
 {% load staticfiles %}
 {% load i18n %}
 {% load bootstrap_tags %}
