+++ valuenetwork/templates/account/logout.html
@@ -1,6 +1,6 @@
 {% extends "work_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 
 {% block extra_head %}
