+++ valuenetwork/templates/account/password_change.html
@@ -1,6 +1,6 @@
 {% extends "account/base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 {% load bootstrap_tags %}
