+++ valuenetwork/templates/account/password_reset_token_fail.html
@@ -1,6 +1,6 @@
 {% extends "work_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 {% load bootstrap_tags %}
