+++ valuenetwork/templates/account/settings.html
@@ -1,6 +1,6 @@
 {% extends "account/base.html" %}
 
-{% load url from future %}
+
 {% load bootstrap_tags %}
 {% load i18n %}
