+++ valuenetwork/templates/account/signup_closed.html
@@ -1,6 +1,6 @@
 {% extends "site_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 {% load bootstrap_tags %}
