+++ valuenetwork/templates/account/work_password_change.html
@@ -1,6 +1,6 @@
 {% extends "work_base.html" %}
 
-{% load url from future %}
+
 {% load i18n %}
 {% load bootstrap_tags %}
