+++ valuenetwork/templates/account/work_settings.html
@@ -1,6 +1,6 @@
 {% extends "work_base.html" %}
 
-{% load url from future %}
+
 {% load bootstrap_tags %}
 {% load i18n %}
