+++ valuenetwork/templates/notification/email_body.txt
@@ -1,4 +1,4 @@
-{% load i18n %}{% load url from future %}{% url "notification_notices" as notices_url %}{% blocktrans %}You have received the following notice from {{ site_name }}:
+{% load i18n %}{% url "notification_notices" as notices_url %}{% blocktrans %}You have received the following notice from {{ site_name }}:
 
 {{ message }}
