+++ valuenetwork/templates/notification/notice_settings.html
@@ -1,7 +1,7 @@
 {% extends "notification/base.html" %}
 
 {% load i18n %}
-{% load url from future %}
+
 
 {% block body_id %}notification-settings{% endblock %}
