+++ valuenetwork/templates/valueaccounting/agent_associations.html
@@ -2,7 +2,7 @@
 
 {% load staticfiles %}
 {% load i18n %}
-{% load url from future %}
+
 {% load bootstrap_tags %}
