+++ valuenetwork/templates/valueaccounting/change_event.html
@@ -1,7 +1,7 @@
 {% extends "site_base.html" %}
 
 {% load i18n %}
-{% load url from future %}
+
 {% load bootstrap_tags %}
 {% load staticfiles %}
