+++ valuenetwork/templates/valueaccounting/order_delete_confirmation.html
@@ -1,7 +1,7 @@
 {% extends "site_base.html" %}
 
 {% load i18n %}
-{% load url from future %}
+
 
 {% block head_title %}{% trans "Confirm deletion of " %}{{ order }}{% endblock %}
