+++ valuenetwork/templates/valueaccounting/unscheduled_time_contributions.html
@@ -2,7 +2,7 @@
 
 {% load staticfiles %}
 {% load i18n %}
-{% load url from future %}
+
 {% load bootstrap_tags %}
