+++ valuenetwork/templates/valueaccounting/value_equation.html
@@ -1,7 +1,7 @@
 {% extends "site_base.html" %}
 
 {% load i18n %}
-{% load url from future %}
+
 {% load bootstrap_tags %}
 
 {% block head_title %}{% trans "Value Time Contributions for " %} {{ project }}{% endblock %}
