+++ work/templates/new_base.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 {% load staticfiles %}
 {% load i18n %}
-{% load url from future %}
+
 
 <html lang="{{ LANGUAGE_CODE }}">
     <head>
