+++ work/templates/work/non_process_logging.html
@@ -2,7 +2,7 @@
 
 {% load staticfiles %}
 {% load i18n %}
-{% load url from future %}
+
 {% load bootstrap_tags %}
